Added support for CV32E40X

+++ tests/magia/mesh/hello_world/src/test.c
@@ -16,10 +16,11 @@ int main(void)
 
     int test = NULL;
 
-    volatile float16 a = -12.0456f;
-    volatile float16 b = 4.0f;
-    volatile float16 c = (a + b);
-    int test_integer   = -123;
+    volatile float a = -12.0456f;
+    volatile float b = 4.0f;
+    volatile float c = a + b;
+
+    int test_integer = -123;
 
     uint32_t *v = (uint32_t *)&c;
     printf("Hello world! %#x\n", *v);
